+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `Var::flat_map_vec` for mapping `Var<Vec<T>>` to `Var<Vec<O>>` where `T` projects a `Var<O>`.
 * Default setting editor reset button is now localizable in the `zng-wgt-settings` l10n resources.
     - Also surface it in `zng::config::settings::editor::reset_button` for use in custom editors.
 * Implement `FromStr` for resolution units.